refact(upgrade): restructure upgrade pkg
