+++ cmd/manager/main.go
@@ -22,7 +22,6 @@ import (
 	"fmt"
 	"github.com/openebs/node-disk-manager/pkg/setup"
 	"github.com/openebs/node-disk-manager/pkg/upgrade"
-	"github.com/openebs/node-disk-manager/pkg/upgrade/preupgrade"
 	"os"
 	"runtime"
 	"sigs.k8s.io/controller-runtime/pkg/client"
@@ -150,6 +149,6 @@ func main() {
 // performUpgrade performs the upgrade operations
 func performUpgrade(client client.Client) error {
 	//TODO: this task should be named for release, not for upgrade steps
-	preUpgradeTask := preupgrade.NewPreUpgradeTask("1.0", "1.1", client)
+	preUpgradeTask := upgrade.NewPreUpgradeTask("1.0", "1.1", client)
 	return upgrade.RunUpgrade(preUpgradeTask)
 }
